Check if the correct PlayerPiece is being dragged

+++ src/java_/game/tile/FloorTile.java
@@ -93,7 +93,7 @@ public class FloorTile extends Tile {
                 '}';
     }
 
-    public int getPathsBits() {
+    public byte getPathsBits() {
         return pathsBits;
     }
